# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Final Project API
  version: 1.0.0
  description: API for managing users, items, categories, and offers.
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        displayName:
          type: string
          example: "JohnDoe"
        email:
          type: string
          example: "V4Z2a@example.com"
      required:
        - id
        - displayName
        - email
    UserCreate:
      type: object
      properties:
        email:
          type: string
          example: "V4Z2a@example.com"
        password:
          type: string
          example: "securePassword123"
        displayName:
          type: string
          example: "JohnDoe"
      required:
        - email
        - password
        - displayName
...
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Electronics"
      required:
        - id
        - name
    CategoryCreate:
      type: object
      properties:
        name:
          type: string
          example: "Electronics"
      required:
        - name
...
    Item:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Sample Item"
        price:
          type: number
          example: 100.00
        categoryId:
          type: integer
          example: 3
        userId:
          type: integer
          example: 5
      required:
        - id
        - title
        - price
        - categoryId
        - userId
    ItemCreate:
      type: object
      properties:
        title:
          type: string
          example: "Sample Item"
        price:
          type: number
          example: 100.00
        categoryId:
          type: integer
          example: 3
        userId:
          type: integer
          example: 5
      required:
        - title
        - price
        - categoryId
        - userId
...
    Offer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        amount:
          type: number
          example: 250.00
        type:
          type: string
          example: "bid"
        status:
          type: string
          example: "pending"
        itemId:
          type: integer
          example: 4
        userId:
          type: integer
          example: 2
      required:
        - id
        - amount
        - type
        - status
        - itemId
        - userId
    OfferCreate:
      type: object
      properties:
        amount:
          type: number
          example: 199.99
        type:
          type: string
          example: "offer"
        status:
          type: string
          example: "pending"
        itemId:
          type: integer
          example: 3
        userId:
          type: integer
          example: 7
      required:
        - amount
        - type
        - status
        - itemId
        - userId
...
    BadRequestError:
      type: object
      properties:
        message:
          type: string
          example: "The request was invalid."
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
...
security:
  - BearerAuth: []
paths:
  /users:
    get:
      summary: List all users
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
    post:
      summary: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
...
  /categories:
    get:
      summary: List all categories
      tags:
        - Categories
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Create a new category
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
...
  /items:
    get:
      summary: List all items
      tags:
        - Items
      responses:
        '200':
          description: Items retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      summary: Create a new item
      tags:
        - Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
...
  /offers:
    get:
      summary: List all offers
      tags:
        - Offers
      responses:
        '200':
          description: Offers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Offer'
    post:
      summary: Create a new offer
      tags:
        - Offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfferCreate'
      responses:
        '201':
          description: Offer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
...